Implement OrderItemService, OrderItemServiceImplementation and updated Spring Security and OrderItem

+++ src/main/java/com/jsp/entity/Order.java
@@ -16,7 +16,7 @@ import java.time.LocalDate;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int orderId;
+    private long orderId;
 
     private LocalDate orderDate;
     private Double totalAmount;
